Lab 5: bind form submission to class method to execute the form
The method can access the form values and the status of the form

Use EventEmitter to let the parent decide what to do with the form. This ensures form reusability.
Describe submit function. here we create a new contact with the data that comes form the input fields on the form.
We return this object via the eventEmitter.