Refactor contact component and model to support list of objects from Firebase
Firebase doesn't support arrays, but lists of objects. Each object has a unique key.
This key needs to be added to the structure of the code. => id

NOTE: a required parameter cannot follow an optional parameter => if id is optional, put it at the end of the constructor parameters!

Remove also the index given to the contact as this one has its own id now