Contact detail: update
To update the data, we can re-use the form from app-contact-form
The form needs to be adjusted so @Input can also process data.

@Input: the form can handle update data (by modifying the parameters of the FormControl for each variable)

Hide the form when it's not needed => toggleEditing
Link de hidden-attribute to the contact-component and refactor the *ngIf of the ocntact-form so it only shows when "editing" = true.
Message to say updated was succesful

The form emits the data to the parent => Link in the contact-detail.component.html and .ts => update functionality